parts 1 and 2

+++ Journal/Journal/View Controllers/EntriesTableViewController.swift
@@ -11,6 +11,8 @@ import CoreData
 
 class EntriesTableViewController: UITableViewController {
     
+    let entryController = EntryController()
+    
     lazy var fetchedResultsController: NSFetchedResultsController<Entry> = {
         let fetchRequest: NSFetchRequest<Entry> = Entry.fetchRequest()
         fetchRequest.sortDescriptors = [NSSortDescriptor(key: "mood", ascending: true),
@@ -25,77 +27,78 @@ class EntriesTableViewController: UITableViewController {
         return frc
     }()
     
-//    
-//    override func viewWillAppear(_ animated: Bool) {
-//        tableView.reloadData()
-//    }
-
+    //
+    //    override func viewWillAppear(_ animated: Bool) {
+    //        tableView.reloadData()
+    //    }
+    
     // MARK: - Table view data source
-
+    
     override func numberOfSections(in tableView: UITableView) -> Int {
         return fetchedResultsController.sections?.count ?? 1
     }
-
+    
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return fetchedResultsController.sections?[section].numberOfObjects ?? 0
     }
-
- 
+    
+    
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         
         guard let cell = tableView.dequeueReusableCell(withIdentifier: EntryTableViewCell.reuseIdentifier, for: indexPath) as? EntryTableViewCell else {
             fatalError("Can't dequeue cell of type \(EntryTableViewCell.reuseIdentifier)")
         }
-
+        
         cell.entry = fetchedResultsController.object(at: indexPath)
         return cell
     }
-
-
+    
+    
     // Override to support editing the table view.
     override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
         if editingStyle == .delete {
             let entry = fetchedResultsController.object(at: indexPath)
-            CoreDataStack.shared.mainContext.delete(entry)
-            do {
-                try CoreDataStack.shared.mainContext.save()
-            } catch {
-                CoreDataStack.shared.mainContext.reset()
-                NSLog("Error saving managed object context: \(error)")
+            
+            entryController.deleteEntryFromServer(entry: entry) { result in
+                guard let _ = try? result.get() else {
+                    return
+                }
+                
+                CoreDataStack.shared.mainContext.delete(entry)
+                do {
+                    try CoreDataStack.shared.mainContext.save()
+                } catch {
+                    CoreDataStack.shared.mainContext.reset()
+                    NSLog("Error saving managed object context: \(error)")
+                }
+                
             }
+            
         }    
     }
-
-
-
-
+    
+    
+    
+    
     // MARK: - Navigation
-
+    
     // In a storyboard-based application, you will often want to do a little preparation before navigation
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
         // Get the new view controller using segue.destination.
         // Pass the selected object to the new view controller.
-
-        switch segue.identifier {
         
-        case "ShowCreateEntrySegue":
-            guard let destinationVC = segue.destination as? CreateEntryViewController else { return }
-            guard let indexPath = tableView.indexPathForSelectedRow else { return }
-            destinationVC.entry = fetchedResultsController.object(at: indexPath)
-            
-        case "ShowViewEntrySegue":
-            guard let destinationVC = segue.destination as? EntryDetailViewController,
-                let indexPath = tableView.indexPathForSelectedRow else { return }
-            
-            destinationVC.entry = fetchedResultsController.object(at: indexPath)
-            
-        default:
-            break
-
+        if segue.identifier == "ShowViewEntrySegue" {
+            if let detailVC = segue.destination as? EntryDetailViewController,
+                let indexPath = tableView.indexPathForSelectedRow {
+                detailVC.entry = fetchedResultsController.object(at: indexPath)
+            } else if segue.identifier == "ShowCreateEntrySegue" {
+                if let navC = segue.destination as? UINavigationController,
+                    let createEntryVC = navC.viewControllers.first as? CreateEntryViewController {
+                    createEntryVC.entryController = entryController
+                }
+            }
         }
     }
-
-
 }
 
 extension EntriesTableViewController: NSFetchedResultsControllerDelegate {
